backend/requirements.txt
Django==2.2.4
djangorestframework==3.10.2
numpy==1.17.0
pycryptodome==3.8.2